+++ src/components/App.js
@@ -1,13 +1,13 @@
 import React from "react";
-import Main from "../page/Main";
+import Main from "./page/Main";
 import {Route, Switch} from "react-router-dom";
-import MyList from "../page/MyList";
-import MoviePage from "../page/MoviePage";
-import AddReview from "../page/AddReview";
-import Player from "../page/Player";
-import NotFoundScreen from "../page/NotFoundScreen";
+import MyList from "./page/MyList";
+import MoviePage from "./page/MoviePage";
+import AddReview from "./page/AddReview";
+import Player from "./page/Player";
+import NotFoundScreen from "./page/NotFoundScreen";
 import PrivateRoute from "./PrivateRoute";
-import AuthScreen from "../page/AuthScreen";
+import AuthScreen from "./page/AuthScreen";
 
 const App = () => {
